+++ engine/glue.h
@@ -2,7 +2,7 @@
 #define _ENGINE_GLUE_COMP_H
 
 #include "node.h"
-NodeRef glue_node(double threshold, double ratio, double attack_ms,
-                  double release_ms, double makeup_gain, double knee_width,
+NodeRef glue_node(float threshold, float ratio, float attack_ms,
+                  float release_ms, float makeup_gain, float knee_width,
                   NodeRef input);
 #endif
